Fix 15 IMPORTANT Severity CODEQL cmp narrow with wide in loop condition (#732)
* fix codeql string_c_utils.c: 430

* fix codeql source_update_cache_utils: 49

* fix workflow_utils.c: 368

* fix file_info_utils.c: 172

* fix swupdate_handler_v2.cpp: 353

* fix workflow_utils.c: 284

* fix hash_utils.c: 107

* fix script_handler.cpp: 369

* fix path_utils.c: 38

* fix d2c_messaging.c: 261 (now 274)

* fix blob_storage_helper.cpp: 79

* fix root_key_utils.c: 216 (now, 220)

* fix swupdate_handler_v2.cpp: 741

* fix workflow_utils.c: 269

+++ src/extensions/step_handlers/script_handler/src/script_handler.cpp
@@ -366,7 +366,7 @@ ADUC_Result ScriptHandlerImpl::PrepareScriptArguments(
 
     Log_Info("Parsing script arguments: %s", arguments);
     argumentList = ADUC::StringUtils::Split(fileArgs, ' ');
-    for (int i = 0; i < argumentList.size(); i++)
+    for (size_t i = 0; i < argumentList.size(); i++)
     {
         const std::string argument = argumentList[i];
         if (!argument.empty())
